test: 💍 Adds tests to GetOrCreatedCachedColor
